Add a docker image for running builds in
This also adds a few scripts to help out with builds. The standard
`make` and `make run` should basically just do the build inside the
container itself.

Commands like qemu end up running outside the container. Not sure if
this is the exact layout/experience you'd like, but hopefully this helps
as a start!